wip: Make PendingSession/ExistingSession queries composable
