+++ src/ai/backend/manager/scheduler/predicates.py
@@ -13,15 +13,13 @@ from ai.backend.common.types import (
     ResourceSlot, SessionTypes,
 )
 
-from ai.backend.manager.models.kernel import recalc_concurrency_used
-
 from ..models import (
     domains, groups, kernels, keypairs,
     keypair_resource_policies,
     query_allowed_sgroups,
     DefaultForUnspecified,
 )
-from . import (
+from .types import (
     SchedulingContext,
     PendingSession,
     PredicateResult,
@@ -90,19 +88,7 @@ async def check_concurrency(
     )
     await db_conn.execute(query)
 
-    async def rollback(
-        db_conn: SAConnection,
-        sched_ctx: SchedulingContext,
-        sess_ctx: PendingSession,
-    ) -> None:
-        # Instead of subtraction, we recalculate the access_key's usage,
-        # because asynchronous container launch failures and agent failures
-        # (especially with multi-node multi-container cluster sessions)
-        # may accumulate up multiple subtractions, resulting in
-        # negative concurrency_occupied values.
-        await recalc_concurrency_used(db_conn, sess_ctx.access_key)
-
-    return PredicateResult(True, failure_cb=rollback)
+    return PredicateResult(True)
 
 
 async def check_dependencies(
